pkgs: Add systemd-udev (to replace later)
Since Network Manager and its dependancies have been removed, the boots
fails on the serial service:

```
[ TIME ] Timed out waiting for device /dev/ttyLF0.
[DEPEND] Dependency failed for Serial Getty on ttyLF0
```

So after some studies, this is due the lack of `systemd-udev` package
which is mandatory for a basic systemd based OS. However this package
size is more than 12Mb! So I'll later remove it and only take necessary
subpackages.

Signed-off-by: Valentin Geffroy <valentin.geffroy@iot.bzh>